requirements.txt
anyio==3.6.1
cffi==1.15.0
click==8.1.3
cryptography==37.0.2
fastapi==0.78.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
pycparser==2.21
pydantic==1.9.1
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.20.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
typing-extensions==4.2.0
uvicorn==0.18.2
